+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # stories
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/web@0.25.0
+
 ## 0.0.1
 
 ### Patch Changes